Replace checked_* arithmetics with normal operators

See merge request nosana-ci/apps/platform/spl!73

+++ programs/nosana-rewards/src/instructions/sync.rs
@@ -24,7 +24,7 @@ pub fn handler(ctx: Context<Sync>) -> Result<()> {
     stats.remove_rewards_account(reward.reflection, reward.xnos);
 
     // re-enter the pool with the current stake
-    let amount: u128 = reward.get_amount(stats.rate) as u128;
+    let amount: u128 = reward.get_amount(stats.rate);
     reward.update(stats.add_rewards_account(stake.xnos, amount), stake.xnos);
 
     // finish
